backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
whitenoise>=6.6
gunicorn>=21.2
python-decouple>=3.8
